requirements.txt
requests==2.31.0
termcolor==2.5.0
pillow==10.0.0
